NEW BRANCH - BIG KLUDGE
Rolled back to last year's version, and did
a few edits to get it to compile.  Needs to
be tested against current libraries.

It is compiling here with the test setup, but
no specifics are known (as many things were
tried which did not succeed, leaving the testbed
Arduino IDE in a non-standard state).

	modified:   yaffa.h

On branch testing-hh-bkf-aa